Fixing that images with 256 in size would break

+++ srclib/lib_image.h
@@ -13,8 +13,8 @@ namespace ImageFormat {
 };
 
 struct ImageData {
-    uint8_t width;
-    uint8_t height;
+    uint16_t width;
+    uint16_t height;
     const unsigned char* data;
     uint16_t transparentColor;
     uint8_t format;
